+++ packages/config/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - refactor: In CJS, .default has to be required
 - refactor: Migrate to TypeScript
 - fix: Do not return root namespace in config.namespaces
+- feat: Throw an error when creating namespace if it already exists, as creating an option does
 
 ## [1.4.0] - 2022-09-01